infrastructure: make 'bot' as standalone django app
