+++ games/urls.py
@@ -1,15 +1,7 @@
-import os
-
 from django.urls import path
-
-from dotenv import load_dotenv, find_dotenv
 
 from games import views
 
-
-load_dotenv(find_dotenv())
-
-stats_bot_token = os.getenv("STATS_BOT_TOKEN_TEST")
 
 app_name = 'games'
 
@@ -19,5 +11,4 @@ urlpatterns = [
     path('add_game/', views.GamesAddView.as_view(), name='add_game'),
     path('add_game_from_bot/', views.GamesAddBotView.as_view(), name='add_game_from_bot'),
     path('game_check_for_bot/', views.GameCheck.as_view(), name='game_check'),
-    path('stats_proceed/' + stats_bot_token, views.stats_proceed_view, name='stats_proceed_from_bot'),
 ]
