+++ telegram_bot/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class TelegramBotConfig(AppConfig):
+    name = 'telegram_bot'
